Move some util functions out of server/app.js
